requirements.txt
absl-py==2.1.0
certifi
charset-normalizer
contourpy==1.2.1
cycler==0.12.1
easydict==1.13
fonttools==4.51.0
grpcio==1.63.0
idna
importlib_metadata==7.1.0
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl-fft
mkl-random
mkl-service==2.4.0
numpy
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow
protobuf==5.26.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions
tzdata==2024.1
urllib3
Werkzeug==3.0.3
zipp==3.18.1
